added composer require packages and also changed the migration from integer to biginteger

+++ src/Database/Migrations/2018_07_10_063753_add_dz_custom_forms_form_type_id_table.php
@@ -13,7 +13,7 @@ class AddDzCustomFormsFormTypeIdTable extends Migration {
      */
     public function up () {
         Schema::table('dz_custom_forms', function (Blueprint $table) {
-            $table->unsignedInteger('form_type_id')->nullable();
+            $table->unsignedBigInteger('form_type_id')->nullable();
             $table->foreign('form_type_id')->references('id')->on('dz_lookup_values');
         });
